+++ cf/data/array/mixin/__init__.py
@@ -3,4 +3,3 @@ from .arraymixin import ArrayMixin
 from .cfamixin import CFAMixin
 from .compressedarraymixin import CompressedArrayMixin
 from .filearraymixin import FileArrayMixin
-from .indexmixin import IndexMixin