+++ cf/data/dask_utils.py
@@ -7,101 +7,12 @@ instance, as would be passed to `dask.array.map_blocks`.
 
 from functools import partial
 
-import dask.array as da
 import numpy as np
-from dask.core import flatten
+from cfdm.data.dask_utils import cfdm_to_memory
 from scipy.ndimage import convolve1d
 
 from ..cfdatetime import dt, dt2rt, rt2dt
-from ..functions import atol as cf_atol
-from ..functions import rtol as cf_rtol
 from ..units import Units
-
-
-def _da_ma_allclose(x, y, masked_equal=True, rtol=None, atol=None):
-    """An effective dask.array.ma.allclose method.
-
-    True if two dask arrays are element-wise equal within a tolerance.
-
-    Equivalent to allclose except that masked values are treated as
-    equal (default) or unequal, depending on the masked_equal
-    argument.
-
-    Define an effective da.ma.allclose method here because one is
-    currently missing in the Dask codebase.
-
-    Note that all default arguments are the same as those provided to
-    the corresponding NumPy method (see the `numpy.ma.allclose` API
-    reference).
-
-    .. versionadded:: 3.14.0
-
-    :Parameters:
-
-        x: a dask array to compare with y
-
-        y: a dask array to compare with x
-
-        masked_equal: `bool`, optional
-            Whether masked values in a and b are considered equal
-            (True) or not (False). They are considered equal by
-            default.
-
-        {{rtol: number, optional}}
-
-        {{atol: number, optional}}
-
-    :Returns:
-
-        `bool`
-            A Boolean value indicating whether or not the two dask
-            arrays are element-wise equal to the given *rtol* and
-            *atol* tolerance.
-
-    """
-    # TODODASK: put in a PR to Dask to request to add as genuine method.
-
-    if rtol is None:
-        rtol = cf_rtol()
-    if atol is None:
-        atol = cf_atol()
-
-    # Must pass rtol=rtol, atol=atol in as kwargs to allclose, rather than it
-    # using those in local scope from the outer function arguments, because
-    # Dask's internal algorithms require these to be set as parameters.
-    def allclose(a_blocks, b_blocks, rtol=rtol, atol=atol):
-        """Run `ma.allclose` across multiple blocks over two arrays."""
-        result = True
-        # Handle scalars, including 0-d arrays, for which a_blocks and
-        # b_blocks will have the corresponding type and hence not be iterable.
-        # With this approach, we avoid inspecting sizes or lengths, and for
-        # the 0-d array blocks the following iteration can be used unchanged
-        # and will only execute once with block sizes as desired of:
-        # (np.array(<int size>),)[0] = array(<int size>). Note
-        # can't check against more general case of collections.abc.Iterable
-        # because a 0-d array is also iterable, but in practice always a list.
-        if not isinstance(a_blocks, list):
-            a_blocks = (a_blocks,)
-        if not isinstance(b_blocks, list):
-            b_blocks = (b_blocks,)
-
-        # Note: If a_blocks or b_blocks has more than one chunk in
-        #       more than one dimension they will comprise a nested
-        #       sequence of sequences, that needs to be flattened so
-        #       that we can safely iterate through the actual numpy
-        #       array elements.
-
-        for a, b in zip(flatten(a_blocks), flatten(b_blocks)):
-            result &= np.ma.allclose(
-                a, b, masked_equal=masked_equal, rtol=rtol, atol=atol
-            )
-
-        return result
-
-    axes = tuple(range(x.ndim))
-    return da.blockwise(
-        allclose, "", x, axes, y, axes, dtype=bool, rtol=rtol, atol=atol
-    )
 
 
 def cf_contains(a, value):
@@ -127,8 +38,8 @@ def cf_contains(a, value):
             value.
 
     """
-    a = cf_asanyarray(a)
-    value = cf_asanyarray(value)
+    a = cfdm_to_memory(a)
+    value = cfdm_to_memory(value)
     return np.array(value in a).reshape((1,) * a.ndim)
 
 
@@ -162,7 +73,7 @@ def cf_convolve1d(a, window=None, axis=-1, origin=0):
             Convolved float array with same shape as input.
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
 
     # Cast to float to ensure that NaNs can be stored
     if a.dtype != float:
@@ -183,38 +94,6 @@ def cf_convolve1d(a, window=None, axis=-1, origin=0):
             c = np.ma.masked_invalid(c)
 
     return c
-
-
-def cf_harden_mask(a):
-    """Harden the mask of a masked `numpy` array.
-
-    Has no effect if the array is not a masked array.
-
-    .. versionadded:: 3.14.0
-
-    .. seealso:: `cf.Data.harden_mask`
-
-    :Parameters:
-
-        a: `numpy.ndarray`
-            The array to have a hardened mask.
-
-    :Returns:
-
-        `numpy.ndarray`
-            The array with hardened mask.
-
-    """
-    a = cf_asanyarray(a)
-    if np.ma.isMA(a):
-        try:
-            a.harden_mask()
-        except AttributeError:
-            # Trap cases when the input array is not a numpy array
-            # (e.g. it might be numpy.ma.masked).
-            pass
-
-    return a
 
 
 def cf_percentile(a, q, axis, method, keepdims=False, mtol=1):
@@ -276,7 +155,7 @@ def cf_percentile(a, q, axis, method, keepdims=False, mtol=1):
     """
     from math import prod
 
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
 
     if np.ma.isMA(a) and not np.ma.is_masked(a):
         # Masked array with no masked elements
@@ -351,142 +230,6 @@ def cf_percentile(a, q, axis, method, keepdims=False, mtol=1):
     return p
 
 
-def cf_soften_mask(a):
-    """Soften the mask of a masked `numpy` array.
-
-    Has no effect if the array is not a masked array.
-
-    .. versionadded:: 3.14.0
-
-    .. seealso:: `cf.Data.soften_mask`
-
-    :Parameters:
-
-        a: `numpy.ndarray`
-            The array to have a softened mask.
-
-    :Returns:
-
-        `numpy.ndarray`
-            The array with softened mask.
-
-    """
-    a = cf_asanyarray(a)
-
-    if np.ma.isMA(a):
-        try:
-            a.soften_mask()
-        except AttributeError:
-            # Trap cases when the input array is not a numpy array
-            # (e.g. it might be numpy.ma.masked).
-            pass
-
-    return a
-
-
-def cf_where(array, condition, x, y, hardmask):
-    """Set elements of *array* from *x* or *y* depending on *condition*.
-
-    The input *array* is not changed in-place.
-
-    See `where` for details on the expected functionality.
-
-    .. note:: This function correctly sets the mask hardness of the
-              output array.
-
-    .. versionadded:: 3.14.0
-
-    .. seealso:: `cf.Data.where`
-
-    :Parameters:
-
-        array: numpy.ndarray
-            The array to be assigned to.
-
-        condition: numpy.ndarray
-            Where False or masked, assign from *y*, otherwise assign
-            from *x*.
-
-        x: numpy.ndarray or `None`
-            *x* and *y* must not both be `None`.
-
-        y: numpy.ndarray or `None`
-            *x* and *y* must not both be `None`.
-
-        hardmask: `bool`
-           Set the mask hardness for a returned masked array. If True
-           then a returned masked array will have a hardened mask, and
-           the mask of the input *array* (if there is one) will be
-           applied to the returned array, in addition to any masked
-           elements arising from assignments from *x* or *y*.
-
-    :Returns:
-
-        `numpy.ndarray`
-            A copy of the input *array* with elements from *y* where
-            *condition* is False or masked, and elements from *x*
-            elsewhere.
-
-    """
-    array = cf_asanyarray(array)
-    condition = cf_asanyarray(condition)
-    if x is not None:
-        x = cf_asanyarray(x)
-
-    if y is not None:
-        y = cf_asanyarray(y)
-
-    mask = None
-
-    if np.ma.isMA(array):
-        # Do a masked where
-        where = np.ma.where
-        if hardmask:
-            mask = array.mask
-    elif np.ma.isMA(x) or np.ma.isMA(y):
-        # Do a masked where
-        where = np.ma.where
-    else:
-        # Do a non-masked where
-        where = np.where
-        hardmask = False
-
-    condition_is_masked = np.ma.isMA(condition)
-    if condition_is_masked:
-        condition = condition.astype(bool)
-
-    if x is not None:
-        # Assign values from x
-        if condition_is_masked:
-            # Replace masked elements of condition with False, so that
-            # masked locations are assigned from array
-            c = condition.filled(False)
-        else:
-            c = condition
-
-        array = where(c, x, array)
-
-    if y is not None:
-        # Assign values from y
-        if condition_is_masked:
-            # Replace masked elements of condition with True, so that
-            # masked locations are assigned from array
-            c = condition.filled(True)
-        else:
-            c = condition
-
-        array = where(c, array, y)
-
-    if hardmask:
-        if mask is not None and mask.any():
-            # Apply the mask from the input array to the result
-            array.mask |= mask
-
-        array.harden_mask()
-
-    return array
-
-
 def _getattr(x, attr):
     return getattr(x, attr, False)
 
@@ -531,7 +274,7 @@ def cf_YMDhms(a, attr):
     array([1, 2])
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
     return _array_getattr(a, attr=attr)
 
 
@@ -564,7 +307,8 @@ def cf_rt2dt(a, units):
      cftime.DatetimeGregorian(2000, 1, 2, 0, 0, 0, 0, has_year_zero=False)]
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
+
     if not units.iscalendartime:
         return rt2dt(a, units_in=units)
 
@@ -619,7 +363,7 @@ def cf_dt2rt(a, units):
     [365 366]
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
     return dt2rt(a, units_out=units, units_in=None)
 
 
@@ -660,7 +404,7 @@ def cf_units(a, from_units, to_units):
     [1000. 2000.]
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
     return Units.conform(
         a, from_units=from_units, to_units=to_units, inplace=False
     )
@@ -684,7 +428,7 @@ def cf_is_masked(a):
             values.
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
     out = np.ma.is_masked(a)
     return np.array(out).reshape((1,) * a.ndim)
 
@@ -717,30 +461,5 @@ def cf_filled(a, fill_value=None):
     [[-999    2    3]]
 
     """
-    a = cf_asanyarray(a)
+    a = cfdm_to_memory(a)
     return np.ma.filled(a, fill_value=fill_value)
-
-
-def cf_asanyarray(a):
-    """Convert to a `numpy` array.
-
-    Only do this if the input *a* has an `__asanyarray__` attribute
-    with value True.
-
-    .. versionadded:: NEXTVERSION
-
-    :Parameters:
-
-        a: array_like
-            The array.
-
-    :Returns:
-
-            The array converted to a `numpy` array, or the input array
-            unchanged if ``a.__asanyarray__`` False.
-
-    """
-    if getattr(a, "__asanyarray__", False):
-        return np.asanyarray(a)
-
-    return a